opt in and bug fix
