+++ Cargo.toml
@@ -8,6 +8,9 @@ documentation = "https://docs.rs/cpal"
 license = "Apache-2.0"
 keywords = ["audio", "sound"]
 
+[profile.release]
+debug = true
+
 [dependencies]
 failure = "0.1.5"
 lazy_static = "1.3"
@@ -16,11 +19,9 @@ num-traits = "0.2.6"
 [dev-dependencies]
 hound = "3.4"
 
-[target.'cfg(any(target_os = "windows" ))'.dependencies]
-asio-sys = { version = "0.1", path = "asio-sys" }
-
 [target.'cfg(target_os = "windows")'.dependencies]
 winapi = { version = "0.3", features = ["audiosessiontypes", "audioclient", "coml2api", "combaseapi", "debug", "devpkey", "handleapi", "ksmedia", "mmdeviceapi", "objbase", "std", "synchapi", "winuser"] }
+asio-sys = { version = "0.1", path = "asio-sys" }
 
 [target.'cfg(any(target_os = "linux", target_os = "dragonfly", target_os = "freebsd", target_os = "openbsd"))'.dependencies]
 alsa-sys = { version = "0.1", path = "alsa-sys" }
